Simplify the DCS IO worker interface
Require a separate worker per redundant iface.
This removes a lot of complexity but one critical issue remains:
when sending transfers (spoofing), it is not possible for the IO workers
to fill in the transfer-ID field automatically, since that requires
having a shared state which we can't provide at the moment.
A possible solution is to implement a dedicated node for storing such
shared state, which the IO workers could query as necessary.
This may prove expensive (in the computational sense).

Note that we can't really implement a dedicated node for serializing transfers
into transport frames because this necessarily requires knowledge of not only
which kind of transfer to use (CAN, Ethernet, etc.), but also its specific
configuration (like MTU size etc.); only the IO workers possess such context.